requirements.txt
ccxt>=4.0.0
pandas>=1.5.0
numpy>=1.21.0
flask>=2.0.0
sqlalchemy>=1.4.0
cryptography>=3.4.0
websockets>=10.0
plotly>=5.0.0
pytest>=8.0
bcrypt>=4.0
pytest-cov>=4.0
botocore
responses>=0.23.0
vcrpy>=4.1.1
hvac>=1.1.0
